Ya hice el punto 1 y 2 del Tp9

+++ ViewModels/ProductoViewModel.cs
@@ -13,6 +13,6 @@ namespace tl2_tp8_2025_NahuelCondori99.ViewModels
         [Display(Name = "Precio unitario")]
         [Required(ErrorMessage = "El precio es obligatorio")]
         [Range(0.01, double.MaxValue, ErrorMessage = "El precio debe ser mayor a 0.")]
-        public decimal Precio {get; set;}
+        public int Precio {get; set;}
     }
 }
